image_captioning_project/requirements.txt
uvicorn==0.20.0
tensorflow==2.11.0
numpy==1.21.5
pandas==1.4.4
keras==2.11.0
fastapi==0.89.1
pillow
jinja2
python_multipart



